MTFormer/requirements.txt
easydict
torch
Pillow
albumentations
torchvision
matplotlib
pandas
openpyxl
einops
scikit-learn